--- app/src/main/java/com/example/mycrodiary/account/FirebaseRef.kt
@@ -1,11 +1,0 @@
-package com.example.mycrodiary.account
-
-import com.google.firebase.Firebase
-import com.google.firebase.database.database
-
-class FirebaseRef {
-    companion object {
-        val database = Firebase.database("https://project-my-crop-default-rtdb.asia-southeast1.firebasedatabase.app")
-        val userInfo = database.getReference("userInfo")
-    }
-}
