--- app/src/main/java/com/example/mycrodiary/account/LoginpageActivity.kt
@@ -1,57 +1,0 @@
-package com.example.mycrodiary.account
-
-import android.content.Intent
-import android.os.Bundle
-import android.util.Log
-import android.widget.Toast
-import androidx.appcompat.app.AppCompatActivity
-import com.example.mycrodiary.databinding.ActivityLoginpageBinding
-import com.example.mycrodiary.main.MainpageActivity
-import com.google.firebase.Firebase
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.auth
-
-class LoginpageActivity : AppCompatActivity() {
-
-    private lateinit var auth: FirebaseAuth
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        super.onStart()
-        // Check if user is signed in (non-null) and update UI accordingly.
-
-
-        auth = Firebase.auth
-
-        val binding = ActivityLoginpageBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        val move_signinpage = Intent(this, SignuppageActivity::class.java)
-        val move_mainpage = Intent(this, MainpageActivity::class.java)
-
-        binding.signBtn.setOnClickListener(){
-            startActivity(move_signinpage)
-        }
-
-        binding.loginBtn.setOnClickListener{
-            val email = binding.inputEmail.text.toString()
-            val pw = binding.inputPw.text.toString()
-
-            if (email.isNullOrBlank() || pw.isNullOrBlank()) {
-                Toast.makeText(this, "Email 혹은 Password를 입력해주세요.",Toast.LENGTH_LONG).show()
-            }
-            else{
-                auth.signInWithEmailAndPassword(email, pw).addOnCompleteListener(this) { task ->
-                    if (task.isSuccessful) {
-                        Log.d("TAG", "signInWithEmail:success")
-                        Toast.makeText(this,"로그인 성공",Toast.LENGTH_LONG).show()
-                        startActivity(move_mainpage)
-                    } else {
-                        Log.d("TAG", "signInWithEmail:failure", task.exception)
-                        Toast.makeText(this, "Email 혹은 Password를 확인하세요.", Toast.LENGTH_LONG,).show()
-                    }
-                }
-            }
-        }
-    }
-}
