--- app/src/main/java/com/example/mycrodiary/account/UserInfo.kt
@@ -1,6 +1,0 @@
-package com.example.mycrodiary.account
-
-data class UserInfo(
-    var email: String? = null,
-    var nickname: String? = null
-)
