+++ app/src/main/java/com/example/mycrodiary/accountutil/FirebaseRef.kt
@@ -1,0 +1,11 @@
+package com.example.mycrodiary.accountutil
+
+import com.google.firebase.Firebase
+import com.google.firebase.database.database
+
+class FirebaseRef {
+    companion object {
+        val database = Firebase.database("https://project-my-crop-default-rtdb.asia-southeast1.firebasedatabase.app")
+        val userInfo = database.getReference("userInfo")
+    }
+}
