+++ app/src/main/java/com/example/mycrodiary/accountutil/UserInfo.kt
@@ -1,0 +1,6 @@
+package com.example.mycrodiary.accountutil
+
+data class UserInfo(
+    var email: String? = null,
+    var nickname: String? = null
+)
