+++ app/src/main/java/com/example/mycrodiary/cropdiaryutils/Cropinfo.kt
@@ -1,0 +1,3 @@
+package com.example.mycrodiary.cropdiaryutils
+
+class Cropinfo(val profile:String, val name:String, val nickname:String, val date:String )
