--- app/src/main/java/com/example/mycrodiary/editcroplist/Cropinfo.kt
@@ -1,3 +1,0 @@
-package com.example.mycrodiary.editcroplist
-
-class Cropinfo(val profile:String, val name:String, val nickname:String, val date:String )
