+++ app/src/main/java/com/example/mycrodiary/settingpage/SettingpageActivity.kt
@@ -1,0 +1,19 @@
+package com.example.mycrodiary.settingpage
+
+import android.os.Bundle
+import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
+import com.example.mycrodiary.R
+import com.example.mycrodiary.databinding.ActivitySettingpageBinding
+
+class SettingpageActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        val binding = ActivitySettingpageBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+    }
+}
