--- app/src/main/res/layout/activity_cropdiary.xml
@@ -1,54 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/main"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".cropdiarypages.CropdiaryActivity">
-
-    <com.google.android.material.appbar.AppBarLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="@color/main_color"
-        android:gravity="center"
-        >
-
-        <TextView
-            android:id="@+id/main_page"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:padding="18dp"
-            android:text="나의 작물"
-            android:textSize="20dp"
-            android:textStyle="bold"
-            android:textColor="@color/black"
-            />
-
-    </com.google.android.material.appbar.AppBarLayout>
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:padding="5dp">
-
-        <ListView
-            android:id="@+id/crop_diary_listview"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            />
-
-        <android.widget.ImageButton
-
-            android:id="@+id/add_diary"
-            android:layout_width="match_parent"
-            android:layout_height="100dp"
-            android:background="@color/button_color1"
-            android:src="@drawable/plusicon"
-            />
-
-    </LinearLayout>
-
-
-</LinearLayout>
