+++ app/src/main/res/layout/activity_mainpage.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".main.MainpageActivity">
+    tools:context=".mainpage.MainpageActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
@@ -46,8 +46,9 @@
                 android:layout_width="120dp"
                 android:layout_height="120dp"
                 android:layout_marginBottom="5dp"
-                android:src="@drawable/tree"
-                android:background="@color/button_color1"/>
+                android:background="@color/button_color1"
+                android:src="@drawable/nutrition_outline"
+                />nutri
 
             <TextView
                 android:layout_width="match_parent"
@@ -62,7 +63,8 @@
                 android:layout_width="120dp"
                 android:layout_height="120dp"
                 android:layout_marginBottom="5dp"
-                android:background="@color/button_color2"/>
+                android:background="@color/button_color2"
+                android:src="@drawable/person_outline"/>
 
             <TextView
                 android:layout_width="match_parent"
@@ -86,7 +88,8 @@
                 android:layout_width="120dp"
                 android:layout_height="120dp"
                 android:layout_marginBottom="5dp"
-                android:background="@color/button_color2"/>
+                android:background="@color/button_color2"
+                android:src="@drawable/leaf_outline"/>
 
             <TextView
                 android:layout_width="match_parent"
@@ -102,7 +105,8 @@
                 android:layout_width="120dp"
                 android:layout_height="120dp"
                 android:layout_marginBottom="5dp"
-                android:background="@color/button_color1"/>
+                android:background="@color/button_color1"
+                android:src="@drawable/settings_outline"/>
 
             <TextView
                 android:layout_width="match_parent"
