+++ app/src/main/res/layout/activity_mypage.xml
@@ -16,6 +16,7 @@
         >
 
         <TextView
+            android:id="@+id/my_page"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:padding="18dp"
@@ -27,49 +28,45 @@
 
     </com.google.android.material.appbar.AppBarLayout>
 
-    <TextView
-        android:id="@+id/nickname"
-        android:layout_width="wrap_content"
+    <LinearLayout
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:layout_marginTop="40dp"
-        android:text="테스트 이름"
-        android:textSize="25dp"
-        />
+        android:orientation="vertical"
+        android:padding="5dp">
 
-    <ImageView
-        android:id="@+id/userimage"
-        android:layout_width="120dp"
-        android:layout_height="120dp"
-        android:layout_marginTop="20dp"
-        android:layout_gravity="center"
-        android:background="@color/bulu_green"
-        />
+        <TextView
+            android:id="@+id/nickname"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center"
+            android:layout_marginTop="40dp"
+            android:text="테스트 이름"
+            android:textSize="25dp"
+            />
 
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="10dp"
-        android:layout_gravity="center"
-        android:text="포인트"
-        android:textSize="20dp"
-        />
+        <ImageView
+            android:id="@+id/userimage"
+            android:layout_width="120dp"
+            android:layout_height="120dp"
+            android:layout_marginTop="20dp"
+            android:layout_gravity="center"
+            android:background="@color/bulu_green"
+            />
 
-    <TextView
-        android:id="@+id/email"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        />
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="10dp"
+            android:layout_gravity="center"
+            android:text="포인트"
+            android:textSize="20dp"
+            />
 
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        />
+    </LinearLayout>
 
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        />
+
+
+
