--- app/src/main/res/layout/activity_setting.xml
@@ -1,60 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/main"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".settingpage.SettingActivity">
-
-    <com.google.android.material.appbar.AppBarLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="@color/main_color"
-        android:gravity="center"
-        >
-
-        <TextView
-            android:id="@+id/main_page"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:padding="18dp"
-            android:text="설정"
-            android:textSize="20dp"
-            android:textStyle="bold"
-            android:textColor="@color/black"
-            />
-
-    </com.google.android.material.appbar.AppBarLayout>
-
-    <android.widget.Button
-        android:id="@+id/password_change"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="10dp"
-        android:text="비밀번호 변경"
-        android:textSize="25dp"
-        android:background="@color/white"/>
-
-    <android.widget.Button
-        android:id="@+id/nickname_change"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="10dp"
-        android:text="닉네임 변경"
-        android:textSize="25dp"
-        android:background="@color/white"
-        />
-
-    <android.widget.Button
-        android:id="@+id/update_information"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="10dp"
-        android:text="업데이트 설정"
-        android:textSize="25dp"
-        android:background="@color/white"
-        />
-
-</LinearLayout>
