+++ app/src/main/res/layout/activity_signuppage.xml
@@ -8,7 +8,7 @@
     android:orientation="vertical"
     android:gravity="center"
     android:background="@color/white"
-    tools:context=".account.SignuppageActivity">
+    tools:context=".accountpage.SignuppageActivity">
 
     <TextView
         android:layout_width="wrap_content"
